Merge r664535 from trunk:
core: Fix address-in-use startup failure on some platforms caused
by attempting to set up an IPv4 listener which overlaps with an 
existing IPv6 listener.

The failure occurred on the second pass of the open-logs hook in
a configuration such as the following:

  Listen 8080
  Listen 0.0.0.0:8081
  Listen [::]:8081

During the first pass, the two port 8081 listen recs were 
adjacent and existing logic prevented binding to 0.0.0.0:8081.
On the second pass, they were not adjacent and we then tried
to bind to 0.0.0.0:8081, leading to failure on some platforms
(seen on SLES 9 and Ubuntu 7.10, not seen on many other Unix-ish
platforms).

Leave a note about other unhandled configurations.

Submitted by: trawick
Reviewed by: jim


git-svn-id: http://svn.apache.org/repos/asf/httpd/httpd/branches/2.2.x@664705 13f79535-47bb-0310-9956-ffa450edef68